Set transactions as double
Here I fixed two bugs: the spliting method was returning an int because of returning types, and also the value column on the bills table was set to decimal, but now it will be a float.

+++ database/migrations/2017_08_20_013624_create_bills_table.php
@@ -16,7 +16,7 @@ class CreateBillsTable extends Migration
         Schema::create('bills', function (Blueprint $table) {
             $table->increments('id');
             $table->text('name');
-            $table->decimal('value', 5, 2);
+            $table->float('value', 8, 2);
             $table->timestamps();
         });
     }
